Support sorting by passing in query param
Specify a sort type (currently supporting 'name' and 'grade'). If we can't find the type, it'll default to name descending
Specifiy a sort dir (currently casts to 1 or -1 if you provide an integer). 1 will sort ascending, -1 will sort descending. If the type is unknown (aka you provided a string), it'll default to name descending
Add default sorting type